fix all screens to pull out the picture from assets correctly
