+++ screens/AboutUsScreen.js
@@ -9,7 +9,7 @@ import {
 
 import Text from "../components/Text";
 import SafeAreaView from "../components/SafeAreaView";
-import logo from "../../assets/logo.png";
+import logo from "../assets/logo.png";
 
 export default AboutUsScreen = ({ navigation }) => {
   return (
